[Util] Add a type interface for controlling hoisting into globals
This adds a type interface that allows
a) Control over when to hoist a value of a particular type into a global
b) Inserting casting logic around the global to set preferred storage
types for the global

One implementation is provided for the ranked tensor type, with the idea
being, for non-byte aligned types, we want to avoid const-eval
generating such types directly as MLIR won't store them packed. This
allows us to keep data packed all the way through the compiler and never
exposes the runtime to a non-byte aligned type for this particular flow.
This will be especially helpful when moving to large arrays of structs.